delete some cruft
