+++ pkg/packaging/prepare.go
@@ -2,9 +2,12 @@ package packaging
 
 import (
 	"fmt"
+	"io"
 	"io/ioutil"
 	"os"
 	"path/filepath"
+	"strings"
+	"text/template"
 
 	"github.com/kolide/kit/fs"
 	"github.com/pkg/errors"
@@ -142,21 +145,23 @@ func prepare(flavor Flavor, platform string, po PackageOptions) (prepareOutput, 
 		}
 	}
 
+	scriptDir, err := ioutil.TempDir("", "scriptDir")
+	if err != nil {
+		return prepOut, err
+	}
+	prepOut.ScriptsRoot = scriptDir
+	postinstallFile, err := os.Create(filepath.Join(scriptDir, "postinstall"))
+	if err != nil {
+		return prepOut, err
+	}
+	if err := postinstallFile.Chmod(0755); err != nil {
+		return prepOut, err
+	}
+	defer postinstallFile.Close()
+
 	// handle postinstalls
 	switch flavor {
 	case LaunchD:
-		scriptDir, err := ioutil.TempDir("", "scriptDir")
-		if err != nil {
-			return prepOut, err
-		}
-		prepOut.ScriptsRoot = scriptDir
-		postinstallFile, err := os.Create(filepath.Join(scriptDir, "postinstall"))
-		if err != nil {
-			return prepOut, err
-		}
-		if err := postinstallFile.Chmod(0755); err != nil {
-			return prepOut, err
-		}
 		postinstallOpts := &postinstallTemplateOptions{
 			LaunchDaemonDirectory: sp.dir,
 			LaunchDaemonName:      so.ServiceName,
@@ -164,6 +169,18 @@ func prepare(flavor Flavor, platform string, po PackageOptions) (prepareOutput, 
 		if err := renderPostinstall(postinstallFile, postinstallOpts); err != nil {
 			return prepOut, err
 		}
+	case SystemD:
+		postInstallLauncherContents := `#!/bin/bash
+set -e
+systemctl daemon-reload
+systemctl enable launcher
+systemctl restart launcher`
+		fmt.Fprintf(postinstallFile, postInstallLauncherContents)
+
+	case Init:
+		postInstallLauncherContents := `#!/bin/bash
+sudo service launcher restart`
+		fmt.Fprintf(postinstallFile, postInstallLauncherContents)
 	}
 
 	return prepOut, nil
@@ -230,8 +247,75 @@ func servicePaths(packageRoot, identifier string, flavor Flavor) (svcPath, error
 		sp.serviceName = fmt.Sprintf("com.%s.launcher", identifier)
 		sp.servicePath = filepath.Join(sp.dir, fmt.Sprintf("%s.plist", sp.serviceName))
 	case SystemD:
+		sp.dir = "/etc/systemd/system"
+		sp.servicePath = "launcher.service"
 	case Init:
+		sp.dir = "/etc/init.d/"
+		sp.servicePath = "launcher"
 	}
 	err := os.MkdirAll(filepath.Join(packageRoot, sp.dir), fs.DirMode)
 	return sp, err
+}
+
+type postinstallTemplateOptions struct {
+	LaunchDaemonDirectory string
+	LaunchDaemonName      string
+}
+
+func renderPostinstall(w io.Writer, options *postinstallTemplateOptions) error {
+	postinstallTemplate := `#!/bin/bash
+
+[[ $3 != "/" ]] && exit 0
+
+/bin/launchctl stop {{.LaunchDaemonName}}
+
+sleep 5
+
+/bin/launchctl unload {{.LaunchDaemonDirectory}}/{{.LaunchDaemonName}}.plist
+/bin/launchctl load {{.LaunchDaemonDirectory}}/{{.LaunchDaemonName}}.plist`
+	t, err := template.New("postinstall").Parse(postinstallTemplate)
+	if err != nil {
+		return errors.Wrap(err, "not able to parse postinstall template")
+	}
+	return t.ExecuteTemplate(w, "postinstall", options)
+}
+
+type newSyslogTemplateOptions struct {
+	LogPath string
+	PidPath string
+}
+
+func renderNewSyslogConfig(w io.Writer, options *newSyslogTemplateOptions) error {
+	syslogTemplate := `# logfilename          [owner:group]    mode count size when  flags [/pid_file] [sig_num]
+{{.LogPath}}               640  3  4000   *   G  {{.PidPath}} 15`
+	t, err := template.New("syslog").Parse(syslogTemplate)
+	if err != nil {
+		return errors.Wrap(err, "not able to parse postinstall template")
+	}
+	return t.ExecuteTemplate(w, "syslog", options)
+}
+
+// grpcServerForHostname returns the gRPC server hostname given a web address
+// that was serving the website itself
+func grpcServerForHostname(hostname string) string {
+	switch hostname {
+	case "localhost:5000":
+		return "localhost:8800"
+	case "master.cloud.kolide.net":
+		return "master-grpc.cloud.kolide.net:443"
+	case "kolide.co", "kolide.com":
+		return "launcher.kolide.co:443"
+	default:
+		if strings.Contains(hostname, ":") {
+			return hostname
+		} else {
+			return fmt.Sprintf("%s:443", hostname)
+		}
+	}
+}
+
+// sanitizeHostname will replace any ":" characters in a given hostname with "-"
+// This is useful because ":" is not a valid character for file paths.
+func sanitizeHostname(hostname string) string {
+	return strings.Replace(hostname, ":", "-", -1)
 }
